Add help command and setup user in `user.ts`

+++ src/handlers/clear.ts
@@ -7,7 +7,13 @@ const clearHandler = async (ctx: HandlerContext) => {
 
   const user = await getUser(ctx);
 
+  if (!user) {
+    ctx.reply("You don't have an account yet. Please, send /start to create one.");
+    return;
+  }
+
   await supabase.from('telegram_users').delete().eq('telegram_id', ctx.from?.id);
+  await supabase.from('user_feed_queue').delete().eq('user_id', user.user_id);
   await adminAuthClient.deleteUser(user?.user_id!);
   ctx.reply('Your account has been deleted. You can start over by sending /start');
 };
